+++ Cineaste/Extension/UIImageView+Loading.swift
@@ -18,7 +18,29 @@ extension UIImageView {
 
         kf.indicatorType = .activity
 
+        // check if a poster with another size is already cached
+        let cacheKey: String
+        if size == .original {
+            cacheKey = Movie.posterUrl(from: posterPath, for: .small)
+                .absoluteString
+        } else {
+            cacheKey = Movie.posterUrl(from: posterPath, for: .original)
+                .absoluteString
+        }
+
+        var cachedPoster: UIImage?
+
+        let cache = ImageCache.default
+
+        // if another poster is already in cache, use this as placeholderImage
+        if cache.isCached(forKey: cacheKey) {
+            cache.retrieveImage(forKey: cacheKey) { result in
+                guard case let .success(value) = result else { return }
+                cachedPoster = value.image
+            }
+        }
+
         let posterUrl = Movie.posterUrl(from: posterPath, for: size)
-        kf.setImage(with: posterUrl, placeholder: UIImage.posterPlaceholder)
+        kf.setImage(with: posterUrl, placeholder: cachedPoster ?? UIImage.posterPlaceholder)
     }
 }